refactor: Post 디렉토리 구성 변경
