+++ src/main/java/org/ahpuh/surf/like/service/LikeService.java
@@ -6,8 +6,8 @@ import org.ahpuh.surf.like.converter.LikeConverter;
 import org.ahpuh.surf.like.dto.response.LikeResponseDto;
 import org.ahpuh.surf.like.entity.Like;
 import org.ahpuh.surf.like.repository.LikeRepository;
-import org.ahpuh.surf.post.entity.Post;
-import org.ahpuh.surf.post.repository.PostRepository;
+import org.ahpuh.surf.post.domain.Post;
+import org.ahpuh.surf.post.domain.repository.PostRepository;
 import org.ahpuh.surf.user.domain.User;
 import org.ahpuh.surf.user.domain.UserRepository;
 import org.springframework.stereotype.Service;
