+++ src/main/java/org/ahpuh/surf/post/service/PostService.java
@@ -7,15 +7,15 @@ import org.ahpuh.surf.category.domain.CategoryRepository;
 import org.ahpuh.surf.category.dto.CategorySimpleDto;
 import org.ahpuh.surf.common.cursor.CursorResult;
 import org.ahpuh.surf.common.exception.EntityExceptionHandler;
-import org.ahpuh.surf.post.converter.PostConverter;
+import org.ahpuh.surf.post.domain.Post;
+import org.ahpuh.surf.post.domain.PostConverter;
+import org.ahpuh.surf.post.domain.repository.PostRepository;
 import org.ahpuh.surf.post.dto.ExploreDto;
 import org.ahpuh.surf.post.dto.PostCountDto;
 import org.ahpuh.surf.post.dto.PostScoreCategoryDto;
 import org.ahpuh.surf.post.dto.RecentPostDto;
 import org.ahpuh.surf.post.dto.request.PostRequestDto;
 import org.ahpuh.surf.post.dto.response.*;
-import org.ahpuh.surf.post.entity.Post;
-import org.ahpuh.surf.post.repository.PostRepository;
 import org.ahpuh.surf.s3.FileStatus;
 import org.ahpuh.surf.s3.S3Service;
 import org.ahpuh.surf.user.domain.User;
