+++ src/test/java/org/ahpuh/surf/integration/post/repository/PostQueryDslTest.java
@@ -5,10 +5,10 @@ import org.ahpuh.surf.category.domain.Category;
 import org.ahpuh.surf.category.domain.CategoryRepository;
 import org.ahpuh.surf.follow.entity.Follow;
 import org.ahpuh.surf.follow.repository.FollowRepository;
+import org.ahpuh.surf.post.domain.Post;
+import org.ahpuh.surf.post.domain.repository.PostRepository;
 import org.ahpuh.surf.post.dto.ExploreDto;
 import org.ahpuh.surf.post.dto.QExploreDto;
-import org.ahpuh.surf.post.entity.Post;
-import org.ahpuh.surf.post.repository.PostRepository;
 import org.ahpuh.surf.user.domain.User;
 import org.ahpuh.surf.user.domain.UserRepository;
 import org.assertj.core.api.Assertions;
